Literate Configuration
This is the first commit on the new branch that if all goes as planned
will provide a literate configuration file using org-mode and babel to
generate the init.el while providing documentation and comments on the
init file.

A lot of the new configuration and setup for this branch comes from
watching the System Crafters youtube Emacs from Scratch channel.