MDX Mode and Org Project capture
- Added a custom derived mode for mdx files and some associated
configuration. It doesn't work quite like I want it to and the lsp for
it doesn't seem to work quite right. Needs more research and work
- Org Project capture is an attempt to better organize project based
org files that don't need to live inside the repository themselves. I
added capture templates for notes, bugs and tasks and bound to
automatically file to the coresponding org templates in
notes/projects/. If the project org file doesn't exist a default one
will be generated using a template and the substitute package that was
install for this purpose.